++factorial.hpp >> implements the definition of the differnt functions to find factorial of a number; ++factorial.cpp >> implementation of the factorial functions - removed from file factorial_of_n.cpp and added to this file
