+++ DSA/Rescursion/factorial.hpp
@@ -1,0 +1,8 @@
+#ifndef FACTORIAL_H
+#define FACTORIAL_H
+#include <iostream>
+
+uint64_t factorial_recursive(int);
+uint64_t factorial_iterative(int);
+
+#endif
